+++ modules/apps/web-experience/asset/asset-publisher-web/src/main/java/com/liferay/asset/publisher/web/AssetPublisherPortletLayoutListener.java
@@ -15,6 +15,7 @@
 package com.liferay.asset.publisher.web;
 
 import com.liferay.asset.publisher.web.util.AssetPublisherUtil;
+import com.liferay.asset.util.impl.AssetUtil;
 import com.liferay.journal.service.JournalArticleLocalService;
 import com.liferay.portal.kernel.model.Layout;
 import com.liferay.portal.kernel.model.LayoutTypePortletConstants;
@@ -27,7 +28,6 @@ import com.liferay.portal.kernel.service.SubscriptionLocalService;
 import com.liferay.portal.kernel.util.PortletKeys;
 import com.liferay.portal.kernel.util.StringPool;
 import com.liferay.portal.kernel.util.UnicodeProperties;
-import com.liferay.portlet.asset.util.AssetUtil;
 
 /**
  * Provides the implementation of <code>PortletLayoutListener</code> (in
@@ -35,7 +35,7 @@ import com.liferay.portlet.asset.util.AssetUtil;
  * email subscriptions can be removed when the Asset Publisher is removed from
  * the page.
  *
- * @author Zsolt Berentey
+ * @author     Zsolt Berentey
  * @deprecated As of 2.0.0, with not direct replacement
  */
 @Deprecated