+++ modules/apps/web-experience/export-import/export-import-test/src/testIntegration/java/com/liferay/exportimport/test/ExportedMissingReferenceExportImportTest.java
@@ -154,8 +154,6 @@ public class ExportedMissingReferenceExportImportTest
 	public void tearDown() throws Exception {
 		_dummyFolderStagedModelRepository.deleteStagedModels(null);
 		_dummyStagedModelRepository.deleteStagedModels(null);
-
-		//super.tearDown();
 	}
 
 	@Test
@@ -164,7 +162,7 @@ public class ExportedMissingReferenceExportImportTest
 			DummyFolderPortletKeys.DUMMY_FOLDER_WITH_MISSING_REFERENCE,
 			DummyFolderWithMissingDummyPortletDataHandler.class);
 
-		long[] layoutIds = new long[] {layout.getLayoutId()};
+		long[] layoutIds = {layout.getLayoutId()};
 
 		exportImportLayouts(layoutIds, getImportParameterMap());
 
@@ -181,7 +179,7 @@ public class ExportedMissingReferenceExportImportTest
 			DummyFolderPortletKeys.DUMMY_FOLDER_WITH_MISSING_REFERENCE,
 			DummyFolderWithMissingLayoutPortletDataHandler.class);
 
-		long[] layoutIds = new long[] {layout.getLayoutId()};
+		long[] layoutIds = {layout.getLayoutId()};
 
 		exportImportLayouts(layoutIds, getImportParameterMap());