+++ modules/util/source-formatter/src/main/java/com/liferay/source/formatter/checks/GradleDependenciesCheck.java
@@ -21,6 +21,7 @@ import com.liferay.portal.kernel.util.StringUtil;
 import com.liferay.portal.kernel.util.Validator;
 import com.liferay.portal.tools.ToolsUtil;
 
+import java.io.File;
 import java.io.Serializable;
 
 import java.util.Comparator;
@@ -31,6 +32,7 @@ import java.util.regex.Pattern;
 
 /**
  * @author Hugo Huijser
+ * @author Peter Shin
  */
 public class GradleDependenciesCheck extends BaseFileCheck {
 
@@ -113,7 +115,8 @@ public class GradleDependenciesCheck extends BaseFileCheck {
 			String configuration = _getConfiguration(dependency);
 
 			if (configuration.equals("compile") &&
-				isModulesApp(absolutePath, _projectPathPrefix, false)) {
+				isModulesApp(absolutePath, _projectPathPrefix, false) &&
+				_hasBNDFile(absolutePath)) {
 
 				dependency = StringUtil.replaceFirst(
 					dependency, "compile", "provided");
@@ -141,8 +144,21 @@ public class GradleDependenciesCheck extends BaseFileCheck {
 		return dependency.substring(0, pos);
 	}
 
+	private boolean _hasBNDFile(String absolutePath) {
+		if (!absolutePath.endsWith("/build.gradle")) {
+			return false;
+		}
+
+		int pos = absolutePath.lastIndexOf(StringPool.SLASH);
+
+		File file = new File(absolutePath.substring(0, pos + 1) + "bnd.bnd");
+
+		return file.exists();
+	}
+
 	private final Pattern _dependenciesPattern = Pattern.compile(
-		"^dependencies \\{(.+?\n)\\}", Pattern.DOTALL | Pattern.MULTILINE);
+		"^dependencies \\{(((?![\\{\\}]).)+?\n)\\}",
+		Pattern.DOTALL | Pattern.MULTILINE);
 	private final Pattern _incorrectGroupNameVersionPattern = Pattern.compile(
 		"(^[^\\s]+)\\s+\"([^:]+?):([^:]+?):([^\"]+?)\"(.*?)", Pattern.DOTALL);
 	private final Pattern _incorrectWhitespacePattern = Pattern.compile(