+++ portal-kernel/src/com/liferay/exportimport/kernel/lar/PortletDataContext.java
@@ -129,6 +129,8 @@ public interface PortletDataContext extends Serializable {
 			Element element, String path, ClassedModel classedModel)
 		throws PortalException;
 
+	public boolean addImportedFriendlyURL(String importedFriendlyURL);
+
 	public void addLocks(Class<?> clazz, String key) throws PortalException;
 
 	public void addLocks(String className, String key, Lock lock);
@@ -339,6 +341,8 @@ public interface PortletDataContext extends Serializable {
 	public Element getImportDataRootElement();
 
 	public Element getImportDataStagedModelElement(StagedModel stagedModel);
+
+	public Set<String> getImportedFriendlyURLs();
 
 	public long[] getLayoutIds();